+++ packages/faucet/src/api/requestparser.ts
@@ -2,7 +2,7 @@ import { isNonNullObject } from "@cosmjs/utils";
 
 import { HttpError } from "./httperror";
 
-export interface CreditRequestBodyDataWithDenom {
+export interface CreditRequestBodyData {
   /** The base denomination */
   readonly denom: string;
   /** The recipient address */
@@ -16,14 +16,6 @@ export interface CreditRequestBodyDataWithTicker {
   readonly address: string;
 }
 
-export type CreditRequestBodyData = CreditRequestBodyDataWithDenom | CreditRequestBodyDataWithTicker;
-
-export function isCreditRequestBodyDataWithDenom(
-  data: CreditRequestBodyData,
-): data is CreditRequestBodyDataWithDenom {
-  return typeof (data as CreditRequestBodyDataWithDenom).denom === "string";
-}
-
 export class RequestParser {
   public static parseCreditBody(body: unknown): CreditRequestBodyData {
     if (!isNonNullObject(body) || Array.isArray(body)) {
@@ -31,6 +23,10 @@ export class RequestParser {
     }
 
     const { address, denom, ticker } = body as any;
+
+    if (typeof ticker !== "undefined") {
+      throw new HttpError(400, "The 'ticker' field was removed in CosmJS 0.23. Please use 'denom' instead.");
+    }
 
     if (typeof address !== "string") {
       throw new HttpError(400, "Property 'address' must be a string.");
@@ -40,29 +36,17 @@ export class RequestParser {
       throw new HttpError(400, "Property 'address' must not be empty.");
     }
 
-    if (
-      (typeof denom !== "string" && typeof ticker !== "string") ||
-      (typeof denom === "string" && typeof ticker === "string")
-    ) {
-      throw new HttpError(400, "Exactly one of properties 'denom' or 'ticker' must be a string");
+    if (typeof denom !== "string") {
+      throw new HttpError(400, "Property 'denom' must be a string.");
     }
 
-    if (typeof ticker === "string" && ticker.length === 0) {
-      throw new HttpError(400, "Property 'ticker' must not be empty.");
-    }
-
-    if (typeof denom === "string" && denom.length === 0) {
+    if (denom.length === 0) {
       throw new HttpError(400, "Property 'denom' must not be empty.");
     }
 
-    return denom
-      ? {
-          address: address,
-          denom: denom,
-        }
-      : {
-          address: address,
-          ticker: ticker,
-        };
+    return {
+      address: address,
+      denom: denom,
+    };
   }
 }